src/requirements.txt
opencv-python==4.4.0.42
PyAutoGUI==0.9.50
scikit-image==0.17.2
selenium==3.141.0
xvfbwrapper==0.2.9
ffmpeg-python==0.2.0
